move components and styles to seperate files
