+++ packages/react-ui-kit/src/Form/Checkbox.tsx
@@ -126,7 +126,6 @@ export const Checkbox: React.FC<CheckboxProps<HTMLInputElement>> = React.forward
       ...wrapperCSS,
     })}
     style={style}
-    onClick={event => event.stopPropagation()}
   >
     <input
       type="checkbox"
